ChineseLiterature/requirements.txt
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.0.2
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==1.9.10
django-jsonfield==1.0.1
django-mysql-ssl==0.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.10
django-toolbelt==0.0.1
entrypoints==0.2.3
et-xmlfile==1.0.1
gspread==3.1.0
gunicorn==19.9.0
httplib2==0.12.0
idna==2.8
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jdcal==1.4
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==4.3.0
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.1
numpy==1.16.0
oauth2client==4.1.3
openpyxl==2.5.14
pandas==0.24.0
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
plotly==2.5.1
prometheus-client==0.4.2
prompt-toolkit==2.0.7
psycopg2==2.7.7
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
Pygments==2.2.0
PyMySQL==0.9.3
pyparsing==2.3.1
python-dateutil==2.7.5
pytube==9.4.0
pytz==2018.9
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
rsa==4.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.11.0
soupsieve==1.7.3
splinter==0.10.0
static3==0.7.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
twstock==1.1.1
urllib3==1.24.1
virtualenv==16.4.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
xlrd==1.2.0
